+++ src/commands/react.ts
@@ -3,6 +3,7 @@ import type { ReactCliOptions } from '../app/args.js'
 import { resolveReactEntryFiles } from '../app/react-entry-files.js'
 import { printReactUsageTree } from '../output/ascii/react.js'
 import { graphToSerializableReactTree } from '../output/json/react.js'
+import type { AnalyzeOptions } from '../types/analyze-options.js'
 import type { ReactUsageFilter } from '../types/react-usage-filter.js'
 import type { ReactUsageGraph } from '../types/react-usage-graph.js'
 import { BaseCommand } from './base.js'
@@ -29,6 +30,13 @@ export class ReactCommand extends BaseCommand<
       cwd: this.options.cwd ?? graph.cwd,
       filter: this.getFilter(),
     })
+  }
+
+  protected getAnalyzeOptions(): AnalyzeOptions {
+    return {
+      ...super.getAnalyzeOptions(),
+      includeBuiltins: this.options.includeBuiltins,
+    }
   }
 
   private getFilter(): ReactUsageFilter {